validation helper fn and fixes
