+++ client/src/Components/Categories/category.scss
@@ -2,10 +2,13 @@
 
 .category-item {
   display: flex;
-  & article {
+  article {
     text-align: center;
+    button::before{
+      content: 'Explore ';
+    }
   }
-  & figure {
+  figure {
     margin: 0;
     display: flex;
     align-items: center;
