+++ client/src/_helpers/index.js
@@ -1,3 +1,4 @@
 export * from './fake-backend';
 export * from './history';
 export * from './store';
+export * from './register-validation';
