+++ client/src/index.scss
@@ -1,0 +1,4 @@
+body {
+    font-family: "Dosis";
+}
+
